giving up on penpot
It currently works for when access digit.eclipsehome.lan:9001 but not when you access it with digit.eclipsehome.lan/penpot.
Seems to not want to use the reverse proxy.

Closet thing I could find was located here, but it wasn't that useful: https://github.com/penpot/penpot/issues/2389
Next commit will remove penpot